+++ tests/projects/dlang/console_with_pkgs/src/main.d
@@ -1,6 +1,14 @@
 import std.stdio;
+import std.datetime;
+import util.log;
+import dateparser;
 
 void main()
 {
-    writeln("hello world!");
+    log = Log(stderrLogger, stdoutLogger(LogLevel.info), fileLogger("log"));
+    log.info("hello xmake");
+
+    assert(parse("2003-09-25") == SysTime(DateTime(2003, 9, 25)));
+    assert(parse("09/25/2003") == SysTime(DateTime(2003, 9, 25)));
+    assert(parse("Sep 2003")   == SysTime(DateTime(2003, 9, 1)));
 }