+++ tests/projects/dlang/console_with_pkgs/xmake.lua
@@ -1,6 +1,13 @@
 add_rules("mode.debug", "mode.release")
 
+add_requires("dub::log 0.4.3", {alias = "log"})
+add_requires("dub::dateparser", {alias = "dateparser"})
+add_requires("dub::emsi_containers", {alias = "emsi_containers"})
+add_requires("dub::stdx-allocator", {alias = "stdx-allocator"})
+add_requires("dub::mir-core", {alias = "mir-core"})
+
 target("test")
     set_kind("binary")
     add_files("src/*.d")
+    add_packages("log", "dateparser", "emsi_containers", "stdx-allocator", "mir-core")